requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.6
django-appconf==1.1.0
django-celery-email==3.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
kombu==5.4.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
